ref(metrics): Parsing docs and tests from Metric to Bucket
