+++ relay-metrics/tests/fixtures/buckets.statsd.txt
@@ -1,0 +1,4 @@
+endpoint.response_time@millisecond:36:49:57:68|d|#route:user_index
+endpoint.hits:4|c|#route:user_index
+endpoint.parallel_requests:42:17:220:25:85|g|#route:user_index
+endpoint.users:3182887624:4267882815|s|#route:user_index
