--- relay-metrics/tests/fixtures/metrics.json
@@ -1,22 +1,0 @@
-[
-  {
-    "name": "d:custom/endpoint.response_time@millisecond",
-    "unit": "millisecond",
-    "value": 57.0,
-    "type": "d",
-    "timestamp": 1615889449,
-    "tags": {
-      "route": "user_index"
-    }
-  },
-  {
-    "name": "c:custom/endpoint.hits@none",
-    "unit": "none",
-    "value": 1.0,
-    "type": "c",
-    "timestamp": 1615889449,
-    "tags": {
-      "route": "user_index"
-    }
-  }
-]
